Remove FrontPageMain and move logic to /pages
